# server/services/swagger/api-docs.yaml
openapi: 3.0.0
info:
  title: SGL API
  description: API para el Sistema de Gestión de Laboratorios (SGL). Este sistema está diseñado para facilitar la gestión de equipos, usuarios, reservas, y mantenimiento de laboratorios.
  version: 1.0.0
  contact:
    name: Angelo G
    email: angelogarcia2021@@gmail.com
    url: https://wa.me/593982718318
servers:
- url: http://localhost:3000/api
  description: API local
paths:
...
  /lab/all:
    get:
      summary: "Obtener todos los laboratorios"
      description: |
        Esta ruta permite obtener todos los laboratorios del sistema con soporte de paginación. Si los laboratorios han sido previamente consultados, se recuperan desde la caché para mejorar el rendimiento.

        **Flujo de obtención:**
        1. El cliente realiza una solicitud para obtener todos los laboratorios, especificando la página y el límite de resultados.
        2. El servidor primero verifica si los datos están almacenados en caché.
        3. Si los datos están en caché, se devuelven directamente al cliente.
        4. Si no están en caché, el servidor consulta la base de datos y luego almacena los resultados en caché.
        5. El servidor devuelve la lista de laboratorios con la información de paginación.

        **Códigos de estado:**
        - **200**: Laboratorios obtenidos exitosamente.
        - **404**: No se encontraron laboratorios.
        - **500**: Error en el servidor al obtener los laboratorios.
      operationId: "getAllLabs"
      tags:
      - "Laboratorios"
      security:
      - BearerAuth: []
      parameters:
      - name: "page"
        in: "query"
        description: "Número de página para paginación (predeterminado es 1)."
        required: false
        schema:
          type: integer
          example: 1
      - name: "limit"
        in: "query"
        description: "Número de resultados por página (predeterminado es 10)."
        required: false
        schema:
          type: integer
          example: 10
      responses:
        "200":
          description: "Laboratorios obtenidos exitosamente."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Laboratorios obtenidos exitosamente."
                  data:
                    type: object
                    properties:
                      totalRecords:
                        type: integer
                        example: 15
                      totalPages:
                        type: integer
                        example: 2
                      currentPage:
                        type: integer
                        example: 1
                      recordsPerPage:
                        type: integer
                        example: 10
                      labs:
                        type: array
                        items:
                          type: object
                          properties:
                            id_lab:
                              type: string
                              example: "1a2b3c4d"
                            lab_name:
                              type: string
                              example: "Laboratorio de Biología"
                            description:
                              type: string
                              example: "Laboratorio especializado en investigaciones biológicas"
                            createdAt:
                              type: string
                              format: date-time
                              example: "2024-12-10T11:02:23.426Z"
                            updatedAt:
                              type: string
                              format: date-time
                              example: "2024-12-10T11:02:23.426Z"
        "404":
          description: "No se encontraron laboratorios."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "No se encontraron laboratorios."
        "500":
          description: "Error al obtener los laboratorios."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "¡Ops! Ha ocurrido un error. Por favor, inténtalo de nuevo."
  /lab/get-by-id/{id}:
    get:
      summary: "Obtener un laboratorio por ID"
      description: |
        Esta ruta permite obtener un laboratorio específico por su ID. El servidor primero intenta obtener el laboratorio desde la base de datos y si no se encuentra, devuelve un mensaje de error.

        **Códigos de estado:**
        - **200**: Laboratorio obtenido exitosamente.
        - **404**: Laboratorio no encontrado.
        - **500**: Error en el servidor al obtener el laboratorio.
      operationId: "getLabById"
      tags:
      - "Laboratorios"
      security:
      - BearerAuth: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        description: "ID del laboratorio a obtener"
        schema:
          type: string
          example: "1a2b3c4d"
      responses:
        "200":
          description: "Laboratorio obtenido exitosamente."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Laboratorio obtenido exitosamente."
                  data:
                    type: object
                    properties:
                      id_lab:
                        type: string
                        example: "1a2b3c4d"
                      lab_name:
                        type: string
                        example: "Laboratorio de Biología"
                      description:
                        type: string
                        example: "Laboratorio especializado en investigaciones biológicas"
                      createdAt:
                        type: string
                        format: date-time
                        example: "2024-12-10T11:02:23.426Z"
                      updatedAt:
                        type: string
                        format: date-time
                        example: "2024-12-10T11:02:23.426Z"
        "404":
          description: "Laboratorio no encontrado."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "Laboratorio no encontrado."
        "500":
          description: "Error al obtener el laboratorio."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "¡Ops! Ha ocurrido un error. Por favor, inténtalo de nuevo."
  /lab/create:
    post:
      tags:
      - Laboratorios
      summary: Crear un laboratorio nuevo
      description: Crea un laboratorio nuevo con los datos proporcionados y valida los campos.
      operationId: createLab
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: El nombre del laboratorio
                  example: "Test con validación 7"
                location:
                  type: string
                  description: La ubicación del laboratorio
                  example: "Para cosas de medicina 22"
                description:
                  type: string
                  description: Descripción del laboratorio
                  example: ""
                id_analyst_fk:
                  type: string
                  description: El ID del analista asignado (UUID)
                  example: "37bd2cab-a7e4-47af-9e51-15ea99db2dd5"
              required:
              - name
              - location
              - id_analyst_fk
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: El nombre del laboratorio
                  example: "Test con validación 7"
                location:
                  type: string
                  description: La ubicación del laboratorio
                  example: "Para cosas de medicina 22"
                description:
                  type: string
                  description: Descripción del laboratorio
                  example: ""
                id_analyst_fk:
                  type: string
                  description: El ID del analista asignado (UUID)
                  example: "37bd2cab-a7e4-47af-9e51-15ea99db2dd5"
              required:
              - name
              - location
              - id_analyst_fk
      responses:
        '201':
          description: Laboratorio creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: ID del nuevo laboratorio
                    example: 123
                  name:
                    type: string
                    description: El nombre del laboratorio
                    example: "Test con validación 7"
                  location:
                    type: string
                    description: La ubicación del laboratorio
                    example: "Para cosas de medicina 22"
                  description:
                    type: string
                    description: Descripción del laboratorio
                    example: ""
                  id_analyst_fk:
                    type: string
                    description: El ID del analista asignado (UUID)
                    example: "37bd2cab-a7e4-47af-9e51-15ea99db2dd5"
        '400':
          description: Solicitud incorrecta (por ejemplo, nombre ya existe, usuario no tiene rol, etc.)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: "Laboratorio ya existe con el mismo nombre."
        '404':
          description: Laboratorio o supervisor no encontrado o no disponible
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "Laboratorio no encontrado o supervisor no disponible."
        '500':
          description: Error interno en el servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Error interno en el servidor."
  /lab/update/{id}:
    put:
      tags:
      - Laboratorios
      summary: Actualizar un laboratorio existente
      description: Actualiza los detalles de un laboratorio existente, validando los campos proporcionados.
      operationId: updateLab
      parameters:
      - name: id
        in: path
        required: true
        description: ID del laboratorio a actualizar
        schema:
          type: string
          example: "123"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: El nombre del laboratorio
                  example: "Test 2 new"
                location:
                  type: string
                  description: La ubicación del laboratorio
                  example: "Ubicac tes ca ddd"
                description:
                  type: string
                  description: Descripción del laboratorio
                  example: "d"
                active:
                  type: boolean
                  description: Estado de actividad del laboratorio
                  example: true
                id_supervisor_fk:
                  type: string
                  description: ID del supervisor del laboratorio (UUID)
                  example: "98bc56b3-331d-4843-be34-1dcf169097a1"
              required:
              - name
              - location
              - active
              - id_supervisor_fk
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: El nombre del laboratorio
                  example: "Test 2 new"
                location:
                  type: string
                  description: La ubicación del laboratorio
                  example: "Ubicac tes ca ddd"
                description:
                  type: string
                  description: Descripción del laboratorio
                  example: "d"
                active:
                  type: boolean
                  description: Estado de actividad del laboratorio
                  example: true
                id_supervisor_fk:
                  type: string
                  description: ID del supervisor del laboratorio (UUID)
                  example: "98bc56b3-331d-4843-be34-1dcf169097a1"
              required:
              - name
              - location
              - active
              - id_supervisor_fk
      responses:
        '200':
          description: Laboratorio actualizado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    description: El nombre del laboratorio
                    example: "Test 2 new"
                  location:
                    type: string
                    description: La ubicación del laboratorio
                    example: "Ubicac tes ca ddd"
                  description:
                    type: string
                    description: Descripción del laboratorio
                    example: "d"
                  active:
                    type: boolean
                    description: Estado de actividad del laboratorio
                    example: true
                  id_supervisor_fk:
                    type: string
                    description: ID del supervisor asignado (UUID)
                    example: "98bc56b3-331d-4843-be34-1dcf169097a1"
        '400':
          description: Solicitud incorrecta (por ejemplo, ID inválido o campos faltantes)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: "El campo 'name' es requerido."
        '404':
          description: Laboratorio no encontrado o supervisor no disponible
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "Laboratorio no encontrado o supervisor no disponible."
        '500':
          description: Error interno en el servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Error interno en el servidor."
  /lab/delete/{id}:
    delete:
      tags:
      - Laboratorios
      summary: Eliminar un laboratorio existente
      description: Elimina un laboratorio existente, validando que no esté en uso.
      operationId: deleteLab
      parameters:
      - name: id
        in: path
        required: true
        description: ID del laboratorio a eliminar
        schema:
          type: string
          example: "123"
      responses:
        '200':
          description: Laboratorio eliminado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Laboratorio eliminado exitosamente."
        '400':
          description: El laboratorio está en funcionamiento o la solicitud es incorrecta
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: "El laboratorio está en funcionamiento."
        '404':
          description: Laboratorio no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "Laboratorio no encontrado."
        '500':
          description: Error interno en el servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Error al eliminar el laboratorio."
...
  /reactive/report/pdf:
    get:
      tags:
      - Reactivos
      summary: Generar un reporte en PDF de los reactivos
      description: Genera un reporte en formato PDF de los reactivos registrados en un rango de fechas especificado.
      operationId: generatePdfReport
      parameters:
      - in: query
        name: startDate
        required: false
        schema:
          type: string
          format: date
        description: Fecha de inicio del rango para el reporte (en formato YYYY-MM-DD).
      - in: query
        name: endDate
        required: false
        schema:
          type: string
          format: date
        description: Fecha de fin del rango para el reporte (en formato YYYY-MM-DD).
      responses:
        '200':
          description: Reporte PDF generado exitosamente
          content:
            application/pdf:
              schema:
                type: string
                format: byte
                example: "base64-encoded-pdf-content"
        '404':
          description: No se encontraron registros de reactivos para generar el reporte
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "No se encontraron registros de reactivos para generar el reporte."
        '500':
          description: Error interno en el servidor al generar el reporte PDF
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Error al generar el reporte PDF de reactivos."
  # REACTIVE CONSUMPTION
...